+++ cmake/CompileOptionSelector.cmake
@@ -8,7 +8,7 @@
 # Copyright (c) 2014-2016 Advanced Institute for Computational Science, RIKEN.
 # All rights reserved.
 #
-# Copyright (c) 2016-2017 Research Institute for Information Technology, Kyushu University.
+# Copyright (c) 2016-2018 Research Institute for Information Technology, Kyushu University.
 # All rights reserved.
 #
 ###################################################################################
@@ -41,10 +41,10 @@ macro (AddOptimizeOption)
     endif()
 
   elseif(CMAKE_CXX_COMPILER_ID STREQUAL "PGI")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fastsse")
-    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -fastsse")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fast -Mipa=fast,inline")
+    set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -fast -Mipa=fast,inline")
     if(enable_fapi)
-      set(CMAKE_Fortran_FLAGS "-O3")
+      set(CMAKE_Fortran_FLAGS "-fast -Mipa=fast,inline")
     endif()
 
   else()
@@ -52,21 +52,6 @@ macro (AddOptimizeOption)
   endif()
 endmacro()
 
-macro (AddSSE)
-  if(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
-    if(CMAKE_SYSTEM_NAME MATCHES "Darwin")
-      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -msse3")
-    else()
-      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native")
-    endif()
-
-  elseif(CMAKE_CXX_COMPILER_ID STREQUAL "Intel")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -xHost")
-
-  elseif(CMAKE_CXX_COMPILER_ID STREQUAL "PGI")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fastsse")
-  endif()
-endmacro()
 
 macro (FreeForm)
   if(CMAKE_Fortran_COMPILER MATCHES ".*frtpx$")